Add WaitForReceipt instead of WaitUntikTxPoolEmpty in Loadbot
